Add files via upload
Write a Java program that performs the following tasks.
a. Prompt the user to enter a sentence and print the number of words in the sentence.
(Example Sentence: The quick brown fox jumps over the lazy dog!)
b. Find and display the longest word in the sentence.
c. Reverse the sentence the user has input and display the modified sentence.
[Hint: String reversedSentence = new StringBuilder(sentence).reverse().toString();]